pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.4",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "reportlab>=4.3.1",
    "streamlit>=1.44.1",
]
